Desktop/Scout Accelerator/scouts-accelerator/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
supabase==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic[email]==2.9.2
pydantic-core==2.23.4
python-dateutil==2.8.2
requests==2.32.3
